fix: app.ts compartment
