+++ src/app/app.ts
@@ -1,25 +1,19 @@
 import {
   Component,
   ElementRef,
-  EffectRef,
-  Injector,
   OnDestroy,
   OnInit,
   inject,
   signal,
   viewChild,
-  effect,
 } from '@angular/core';
 import { CanvasComponent } from '../ui/canvas/canvas/canvas.component';
 import { ScriptComponent } from '../ui/script/script/script.component';
 import { TimelineComponent } from '../ui/timeline/timeline/timeline.component';
-import { TimelineActions } from '../ui/timeline/helpers/timeline.actions';
 import { AboutWindowComponent } from '../ui/dialogs/about-window/about-window.component';
 import { SettingsComponent } from '../ui/dialogs/settings/settings.component';
 import { ExportProgressComponent } from '../ui/export-progress/export-progress.component';
 import { ExportSettingsComponent } from '../ui/export-settings/export-settings.component';
-import { AppStore } from 'src/data';
-import { SbdService } from './app.sbd.service';
 import { ThemeService } from '../services/theme.service';
 import { SaveService } from '../services/save.service';
 import { ExportIpcService } from '../services/export-ipc.service';
@@ -27,11 +21,8 @@ import { WindowScalingService } from '../services/window-scaling.service';
 import { UndoRedoService } from '../services/undo-redo.service';
 import { PlaybackService } from '../services/playback.service';
 import { WebToolbarComponent } from '../ui/web-toolbar/web-toolbar.component';
-import { appSettings } from 'src/settings-loader';
 import { Capacitor } from '@capacitor/core';
-import { NativeToolbarService, type ThemeId } from '../services/native-toolbar.service';
-import { AppShortcutsService } from 'src/services';
-import { PlatformFileService } from '../services/platform-file.service';
+import { AppShortcutsService, IosIntegrationService } from 'src/services';
 
 @Component({
   selector: 'app-root',
@@ -59,7 +50,6 @@ export class App implements OnInit, OnDestroy {
   protected readonly saveService = inject(SaveService);
   protected readonly dialogMode = signal<'about' | 'settings' | null>(null);
   private readonly canvas = viewChild(CanvasComponent);
-  private readonly sbd = inject(SbdService);
   private readonly el = inject(ElementRef);
   private readonly themeService = inject(ThemeService);
   private readonly windowScalingService = inject(WindowScalingService);
@@ -70,43 +60,24 @@ export class App implements OnInit, OnDestroy {
   protected readonly useSafeArea = !this.isElectron;
   private readonly undoRedo = inject(UndoRedoService);
   private readonly playback = inject(PlaybackService);
-  private readonly nativeToolbar = inject(NativeToolbarService);
-  private readonly platformFile = inject(PlatformFileService);
-  private readonly injector = inject(Injector);
-  private store = inject(AppStore);
-  private actions = inject(TimelineActions);
-  private settings = appSettings;
+  private readonly iosIntegration = inject(IosIntegrationService);
   private readonly shortcuts = inject(AppShortcutsService);
   private removeThemeListener?: () => void;
   private removeShortcutListener?: () => void | undefined;
-  private removeNativeMenuListener?: () => void;
   private removeWindowScalingListener?: () => void;
   private removeExportIpcListeners?: () => void;
-  private iosAutosaveTimer: ReturnType<typeof setInterval> | null = null;
-  private iosSaveInProgress = false;
-  private iosHasActiveDocument = false;
-  private iosInitialSavePrompted = false;
-  private iosInitialSaveEffectRef?: EffectRef;
-
-  constructor() {
-    effect(() => {
-      if (this.isIos) {
-        this.nativeToolbar.setTitle(this.title());
-        this.nativeToolbar.configureMenu(this.themeService.currentTheme() as ThemeId);
-      }
-    });
-  }
 
   ngOnInit(): void {
     this.removeThemeListener = this.themeService.initTheme();
 
-    if (this.isIos) {
-      void this.nativeToolbar.onMenuAction((actionId) => {
-        void this.handleNativeMenuAction(actionId);
-      }).then((removeListener) => {
-        this.removeNativeMenuListener = removeListener;
-      });
-    }
+    this.iosIntegration.init(
+      {
+        flushCurrentBoardState: (force) => this.canvas()?.flushCurrentBoardState(force),
+        prepareForProjectLoad: () => this.canvas()?.prepareForProjectLoad(),
+      },
+      (dialog) => this.dialogMode.set(dialog),
+      () => this.title()
+    );
 
     // Check if this window was opened as a dialog by the main process
     const params = new URLSearchParams(window.location.search);
@@ -117,10 +88,6 @@ export class App implements OnInit, OnDestroy {
     }
 
     this.saveService.init();
-
-    if (this.isIos) {
-      this.initIosSaveFlow();
-    }
 
     this.removeExportIpcListeners = this.exportIpc.init();
 
@@ -227,207 +194,14 @@ export class App implements OnInit, OnDestroy {
 
   ngOnDestroy(): void {
     this.saveService.destroy();
+    this.iosIntegration.destroy();
     this.removeThemeListener?.();
     this.removeShortcutListener?.();
-    this.removeNativeMenuListener?.();
     this.removeWindowScalingListener?.();
     this.removeExportIpcListeners?.();
-    this.iosInitialSaveEffectRef?.destroy();
-    if (this.iosAutosaveTimer) {
-      clearInterval(this.iosAutosaveTimer);
-      this.iosAutosaveTimer = null;
-    }
-  }
-
-  private getIosSavingSettings(): {
-    autosave: boolean;
-    autosaveDurationMs: number;
-    initialSave: boolean;
-  } {
-    const saving = this.settings.saving;
-    const resolvedAutosave = saving?.autosave ?? this.settings.autosave ?? true;
-    const resolvedAutosaveDuration =
-      saving?.autosaveDuration ?? this.settings.autosaveDuration ?? 300_000;
-    const autosaveDurationMs =
-      typeof resolvedAutosaveDuration === 'number' && resolvedAutosaveDuration > 0
-        ? resolvedAutosaveDuration
-        : 300_000;
-
-    return {
-      autosave: resolvedAutosave,
-      autosaveDurationMs,
-      initialSave: saving?.initialSave ?? true,
-    };
-  }
-
-  private initIosSaveFlow(): void {
-    this.iosInitialSavePrompted = false;
-
-    this.iosInitialSaveEffectRef?.destroy();
-    this.iosInitialSaveEffectRef = effect(
-      () => {
-        const hasUndoHistory = this.undoRedo.canUndo();
-
-        if (!hasUndoHistory) {
-          this.iosInitialSavePrompted = false;
-          return;
-        }
-
-        if (!this.getIosSavingSettings().initialSave) {
-          return;
-        }
-
-        if (this.iosInitialSavePrompted || this.iosHasActiveDocument || this.iosSaveInProgress) {
-          return;
-        }
-
-        this.iosInitialSavePrompted = true;
-        void this.triggerIosSave(false);
-      },
-      { injector: this.injector },
-    );
-
-    const { autosave, autosaveDurationMs } = this.getIosSavingSettings();
-    if (autosave) {
-      this.iosAutosaveTimer = setInterval(() => {
-        if (!this.iosHasActiveDocument || this.iosSaveInProgress) {
-          return;
-        }
-        void this.triggerIosSave(false, false, true);
-      }, autosaveDurationMs);
-    }
-  }
-
-  private async handleNativeMenuAction(actionId: string): Promise<void> {
-    switch (actionId) {
-      case 'app.about':
-        this.openMobileDialog('about');
-        return;
-      case 'app.settings':
-        this.openMobileDialog('settings');
-        return;
-      case 'file.save':
-        await this.triggerIosSave(false);
-        return;
-      case 'file.saveAs':
-        await this.triggerIosSave(true);
-        return;
-      case 'file.load':
-        await this.triggerIosLoad();
-        return;
-      case 'file.export':
-        this.triggerIosExport();
-        return;
-      case 'edit.undo':
-        this.undoRedo.triggerUndo();
-        return;
-      case 'edit.redo':
-        this.undoRedo.triggerRedo();
-        return;
-      default:
-        if (actionId.startsWith('theme.')) {
-          const [, theme] = actionId.split('.') as [string, ThemeId];
-          this.themeService.applyTheme(theme);
-        }
-    }
-  }
-
-  private openMobileDialog(dialog: 'about' | 'settings'): void {
-    this.exportIpc.onSettingsCancel();
-    this.dialogMode.set(dialog);
   }
 
   protected closeDialog(): void {
     this.dialogMode.set(null);
-  }
-
-  private async triggerIosSave(
-    promptForName: boolean,
-    showToast = true,
-    isAutosave = false,
-  ): Promise<boolean> {
-    if (this.iosSaveInProgress) {
-      return false;
-    }
-
-    this.iosSaveInProgress = true;
-
-    try {
-      const isModernWeb = !window.quickboard && 'showSaveFilePicker' in window;
-
-      if (promptForName && !isModernWeb && !this.isIos) {
-        const newName = window.prompt(
-          'Enter file name without extension:',
-          this.exportIpc.defaultPrefix() || 'project',
-        );
-        if (newName) {
-          this.exportIpc.setProjectName(newName);
-        } else {
-          return false;
-        }
-      }
-
-      this.canvas()?.flushCurrentBoardState(true);
-
-      const zipData = await this.sbd.buildSbdZip();
-      const fileName = `${this.exportIpc.defaultPrefix() || 'project'}.sbd`;
-      const saved = await this.platformFile.saveFile(zipData, fileName, undefined, promptForName);
-
-      if (!saved) {
-        if (!isAutosave) {
-          this.iosInitialSavePrompted = false;
-        }
-        return false;
-      }
-
-      this.iosHasActiveDocument = true;
-
-      if (showToast) {
-        this.saveService.saveStatus.set('Saved!');
-        setTimeout(() => {
-          if (this.saveService.saveStatus() === 'Saved!') {
-            this.saveService.saveStatus.set(null);
-          }
-        }, 2000);
-      }
-
-      return true;
-    } catch (error) {
-      console.error('Save failed from native iPad menu', error);
-      window.alert(`Failed to save file: ${error instanceof Error ? error.message : String(error)}`);
-      return false;
-    } finally {
-      this.iosSaveInProgress = false;
-    }
-  }
-
-  private async triggerIosLoad(): Promise<void> {
-    try {
-      this.canvas()?.prepareForProjectLoad();
-      const result = await this.platformFile.pickAndReadFile('.sbd');
-      if (!result) {
-        return;
-      }
-
-      await this.sbd.loadSbdZip(result.data);
-      this.undoRedo.clear();
-
-      const stem = result.name.replace(/\.[^.]+$/, '');
-      if (stem) {
-        this.exportIpc.setProjectName(stem);
-      }
-
-      this.iosHasActiveDocument = true;
-      this.iosInitialSavePrompted = true;
-    } catch (error) {
-      console.error('Load failed from native iPad menu', error);
-      window.alert(`Failed to load file: ${error instanceof Error ? error.message : String(error)}`);
-    }
-  }
-
-  private triggerIosExport(): void {
-    this.dialogMode.set(null);
-    this.exportIpc.settingsBoardCount.set(this.store.boards().length);
-    this.exportIpc.settingsVisible.set(true);
   }
 }
